Basic player list and expanding player domain object.
Added a simple player list page, that shows the player name, as well as
full name and group. The name is a link that can be clicked to go to the
player page, which hasn't been done yet.

Also added a full name and a group to the player domain object. These
properties are to be used for specifying the full name of a player as
well as which group the player belongs to. This can be a logical group
such as "sales vs. dev" or regional such as "Poland vs. Denmark".