Introduced a DatabaseManager for handling evolution of database schema.
Now that we don't have Hibernate any longer, we need a way to handle
schema changes. I have introduced a database manager, that handles
evolution of a database by adding a version table to the database and
upgrade it based on SQL scripts if needed.

See more in the DatabaseManager class for details.