Now requires login to add/edit.
Added a simple login mechanism the enables the adding of games and
editing of players. Everyone can create themselves as users. Later it
should be possible for a user to link to a player (claim it). There
should also be different roles for users, so an admin or moderator can
keep the peace (by banning users).