Upgrade Twitter Bootstrap and FontAwesome.
This required a change to the lists with a pop-over associated with a TD. There is a long running issue[1] with Bootstrap about popover on table cells. That means that currently we have to have a div inside each TD that needs popover. It also means that we don't react to mouseover on the entire TD because of padding. If they resolve issue we can revert back to no div.

[1]: https://github.com/twitter/bootstrap/issues/5687