Prepare switching the Java backend with a Go backend.
The reason is that if we want to use Google App Engine, we must play by their rules. That means that we are not guaranteed that a server instance is launched when a user tries to hit the server. With Java and even a simple Servlet based approach, a cold launch of the application is ~5-15 seconds. If we use Go, the initial request for a cold boot is <500ms, excluding whatever the request actually does (Datastore access etc.)

Go is not as widespread as Java, or even Python, but it is optimized towards serving the backend. Disadvantage is that there are not as many mature frameworks for it, as there is for Java. There are not Objectify framework to make working with the datastore easy.

I will see how it goes implementing the backend using Go, and evaluate along the way.